Add Server-TxClock, tx-status and POST.
Several features have been added to the server. Responses include Server-TxClock so that the client can implement a logical clock. The client can use /tx-status to check the outcome of a past write. And the client can POST to a row to get create semantics. Also, for consistency with just about everything else, the field in a batch write is now `value` instead of `obj`.